feat(notification): add notification model
