+++ src/app/shared/models/notification.model.ts
@@ -1,0 +1,8 @@
+export interface Notification {
+  id: number;
+  type: 'success' | 'error' | 'warning' | 'info' | 'primary';
+  title: string;
+  message: string;
+  duration: number;
+  visible: boolean;
+}
